refactor: reuse components
