+++ src/components/client/edit-account-dialog.tsx
@@ -1,45 +1,13 @@
 "use client";
 
-import { useState } from "react";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
-import { z } from "zod";
-import { AccountStatus, AccountType } from "@/lib/api";
 import { useAgent, Account } from "@/contexts/agent-context";
-
-import { Button } from "@/components/ui/button";
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-  DialogTrigger,
-} from "@/components/ui/dialog";
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "@/components/ui/form";
-import { Input } from "@/components/ui/input";
 import { Edit } from "lucide-react";
-
-// Define the form schema
-const accountFormSchema = z.object({
-  clientId: z.string().min(1, "Client ID is required"),
-  accountType: z.string().min(1, "Account type is required"),
-  accountStatus: z.string().min(1, "Account status is required"),
-  openingDate: z.string().min(1, "Opening date is required"),
-  initialDeposit: z.coerce.number().min(0, "Initial deposit must be a positive number"),
-  currency: z.string().min(1, "Currency is required"),
-  branchId: z.string().min(1, "Branch ID is required"),
-});
-
-type AccountFormValues = z.infer<typeof accountFormSchema>;
+import { Button } from "@/components/ui/button";
+import { FormDialog } from "@/components/forms/form-dialog";
+import { AccountFormFields } from "@/components/forms/account-form-fields";
+import { accountFormSchema, AccountFormValues } from "@/lib/schemas/account-schema";
 
 interface EditAccountDialogProps {
   account: Account;
@@ -47,7 +15,6 @@ interface EditAccountDialogProps {
 }
 
 export function EditAccountDialog({ account, clientName }: Readonly<EditAccountDialogProps>) {
-  const [open, setOpen] = useState(false);
   const { updateAccount, loading } = useAgent();
 
   const form = useForm<AccountFormValues>({
@@ -65,153 +32,23 @@ export function EditAccountDialog({ account, clientName }: Readonly<EditAccountD
 
   async function onSubmit(data: AccountFormValues) {
     await updateAccount(account.id, data);
-    setOpen(false);
   }
 
   return (
-    <Dialog open={open} onOpenChange={setOpen}>
-      <DialogTrigger asChild>
+    <FormDialog
+      title="Edit Account"
+      description={`Edit account details for ${clientName}`}
+      trigger={
         <Button variant="ghost" size="icon" className="h-7 w-7">
           <Edit className="h-3.5 w-3.5" />
         </Button>
-      </DialogTrigger>
-      <DialogContent className="max-h-[90vh] overflow-y-auto sm:max-w-[500px]">
-        <DialogHeader>
-          <DialogTitle className="text-center text-xl">Edit Account</DialogTitle>
-          <DialogDescription className="text-center">
-            Edit account details for {clientName}
-          </DialogDescription>
-        </DialogHeader>
-        <Form {...form}>
-          <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-6">
-            <div className="space-y-4">
-              <div className="rounded-md border p-4">
-                <h3 className="mb-3 text-sm font-medium">Account Information</h3>
-                <div className="grid gap-4">
-                  <FormField
-                    control={form.control}
-                    name="accountType"
-                    render={({ field }) => (
-                      <FormItem>
-                        <FormLabel>Account Type</FormLabel>
-                        <FormControl>
-                          <select
-                            className="flex h-10 w-full rounded-md border border-input bg-background px-3 py-2 text-sm ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50"
-                            {...field}
-                          >
-                            <option value={AccountType.SAVINGS}>Savings</option>
-                            <option value={AccountType.CHECKING}>Checking</option>
-                            <option value={AccountType.BUSINESS}>Business</option>
-                          </select>
-                        </FormControl>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-                  <FormField
-                    control={form.control}
-                    name="accountStatus"
-                    render={({ field }) => (
-                      <FormItem>
-                        <FormLabel>Account Status</FormLabel>
-                        <FormControl>
-                          <select
-                            className="flex h-10 w-full rounded-md border border-input bg-background px-3 py-2 text-sm ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50"
-                            {...field}
-                          >
-                            <option value={AccountStatus.ACTIVE}>Active</option>
-                            <option value={AccountStatus.INACTIVE}>Inactive</option>
-                            <option value={AccountStatus.CLOSED}>Closed</option>
-                            <option value={AccountStatus.PENDING}>Pending</option>
-                          </select>
-                        </FormControl>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-                  <FormField
-                    control={form.control}
-                    name="openingDate"
-                    render={({ field }) => (
-                      <FormItem>
-                        <FormLabel>Opening Date</FormLabel>
-                        <FormControl>
-                          <Input type="date" {...field} />
-                        </FormControl>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-                  <FormField
-                    control={form.control}
-                    name="initialDeposit"
-                    render={({ field }) => (
-                      <FormItem>
-                        <FormLabel>Initial Deposit</FormLabel>
-                        <FormControl>
-                          <Input type="number" step="0.01" min="0" {...field} />
-                        </FormControl>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-                  <FormField
-                    control={form.control}
-                    name="currency"
-                    render={({ field }) => (
-                      <FormItem>
-                        <FormLabel>Currency</FormLabel>
-                        <FormControl>
-                          <select
-                            className="flex h-10 w-full rounded-md border border-input bg-background px-3 py-2 text-sm ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50"
-                            {...field}
-                          >
-                            <option value="SGD">SGD - Singapore Dollar</option>
-                            <option value="USD">USD - US Dollar</option>
-                            <option value="EUR">EUR - Euro</option>
-                            <option value="GBP">GBP - British Pound</option>
-                          </select>
-                        </FormControl>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-                  <FormField
-                    control={form.control}
-                    name="branchId"
-                    render={({ field }) => (
-                      <FormItem>
-                        <FormLabel>Branch</FormLabel>
-                        <FormControl>
-                          <select
-                            className="flex h-10 w-full rounded-md border border-input bg-background px-3 py-2 text-sm ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50"
-                            {...field}
-                          >
-                            <option value="SG001">Singapore - Main Branch</option>
-                            <option value="SG002">Singapore - Orchard Road</option>
-                            <option value="SG003">Singapore - Changi</option>
-                            <option value="MY001">Malaysia - Kuala Lumpur</option>
-                          </select>
-                        </FormControl>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-                </div>
-              </div>
-            </div>
-
-            <DialogFooter className="gap-2 sm:gap-0">
-              <Button type="button" variant="outline" onClick={() => setOpen(false)}>
-                Cancel
-              </Button>
-              <Button type="submit" disabled={loading}>
-                {loading ? "Saving..." : "Save Changes"}
-              </Button>
-            </DialogFooter>
-          </form>
-        </Form>
-      </DialogContent>
-    </Dialog>
+      }
+      form={form}
+      onSubmit={onSubmit}
+      loading={loading}
+      submitLabel="Save Changes"
+    >
+      <AccountFormFields form={form} />
+    </FormDialog>
   );
 }
