+++ src/components/forms/account-form-fields.tsx
@@ -1,0 +1,179 @@
+"use client";
+
+import { AccountStatus, AccountType } from "@/lib/api";
+import { UseFormReturn } from "react-hook-form";
+import { AccountFormValues } from "@/lib/schemas/account-schema";
+
+import {
+  FormControl,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+} from "@/components/ui/form";
+import { Input } from "@/components/ui/input";
+
+interface AccountFormFieldsProps {
+  form: UseFormReturn<AccountFormValues>;
+  showClientSelection?: boolean;
+  clients?: Array<{ id: string; firstName: string; lastName: string; email: string }>;
+  onClientChange?: (clientId: string) => void;
+}
+
+export function AccountFormFields({
+  form,
+  showClientSelection = false,
+  clients = [],
+  onClientChange,
+}: AccountFormFieldsProps) {
+  return (
+    <div className="space-y-4">
+      {showClientSelection && (
+        <div className="rounded-md border p-4">
+          <h3 className="mb-3 text-sm font-medium">Client Selection</h3>
+          <FormField
+            control={form.control}
+            name="clientId"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel>Select Client</FormLabel>
+                <FormControl>
+                  <select
+                    className="flex h-10 w-full rounded-md border border-input bg-background px-3 py-2 text-sm ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50"
+                    value={field.value}
+                    onChange={(e) => {
+                      field.onChange(e);
+                      onClientChange?.(e.target.value);
+                    }}
+                  >
+                    <option value="">-- Select a client --</option>
+                    {clients.map((client) => (
+                      <option key={client.id} value={client.id}>
+                        {client.firstName} {client.lastName} ({client.email})
+                      </option>
+                    ))}
+                  </select>
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+        </div>
+      )}
+      <div className="rounded-md border p-4">
+        <h3 className="mb-3 text-sm font-medium">Account Information</h3>
+        <div className="grid gap-4">
+          <FormField
+            control={form.control}
+            name="accountType"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel>Account Type</FormLabel>
+                <FormControl>
+                  <select
+                    className="flex h-10 w-full rounded-md border border-input bg-background px-3 py-2 text-sm ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50"
+                    {...field}
+                  >
+                    <option value={AccountType.SAVINGS}>Savings</option>
+                    <option value={AccountType.CHECKING}>Checking</option>
+                    <option value={AccountType.BUSINESS}>Business</option>
+                  </select>
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+          <FormField
+            control={form.control}
+            name="accountStatus"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel>Account Status</FormLabel>
+                <FormControl>
+                  <select
+                    className="flex h-10 w-full rounded-md border border-input bg-background px-3 py-2 text-sm ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50"
+                    {...field}
+                  >
+                    <option value={AccountStatus.ACTIVE}>Active</option>
+                    <option value={AccountStatus.INACTIVE}>Inactive</option>
+                    <option value={AccountStatus.CLOSED}>Closed</option>
+                    <option value={AccountStatus.PENDING}>Pending</option>
+                  </select>
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+          <FormField
+            control={form.control}
+            name="openingDate"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel>Opening Date</FormLabel>
+                <FormControl>
+                  <Input type="date" {...field} />
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+          <FormField
+            control={form.control}
+            name="initialDeposit"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel>Initial Deposit</FormLabel>
+                <FormControl>
+                  <Input type="number" step="0.01" min="0" {...field} />
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+          <FormField
+            control={form.control}
+            name="currency"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel>Currency</FormLabel>
+                <FormControl>
+                  <select
+                    className="flex h-10 w-full rounded-md border border-input bg-background px-3 py-2 text-sm ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50"
+                    {...field}
+                  >
+                    <option value="SGD">SGD - Singapore Dollar</option>
+                    <option value="USD">USD - US Dollar</option>
+                    <option value="EUR">EUR - Euro</option>
+                    <option value="GBP">GBP - British Pound</option>
+                  </select>
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+          <FormField
+            control={form.control}
+            name="branchId"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel>Branch</FormLabel>
+                <FormControl>
+                  <select
+                    className="flex h-10 w-full rounded-md border border-input bg-background px-3 py-2 text-sm ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50"
+                    {...field}
+                  >
+                    <option value="SG001">Singapore - Main Branch</option>
+                    <option value="SG002">Singapore - Orchard Road</option>
+                    <option value="SG003">Singapore - Changi</option>
+                    <option value="MY001">Malaysia - Kuala Lumpur</option>
+                  </select>
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+        </div>
+      </div>
+    </div>
+  );
+}
